Refactoring and renaming of Neuroglancer component
